promotional-analytics/deployment/requirements.txt
streamlit==1.32.0
pandas==2.2.1
numpy==1.26.4
google-cloud-storage==2.16.0
google-cloud-aiplatform==1.49.0
